Refactor DB interface
