+++ cmd/manager/v1alpha3/main.go
@@ -6,10 +6,12 @@ import (
 	"flag"
 	"fmt"
 	"net"
+	"os"
 
 	health_pb "github.com/kubeflow/katib/pkg/apis/manager/health"
 	api_pb "github.com/kubeflow/katib/pkg/apis/manager/v1alpha3"
-	kdb "github.com/kubeflow/katib/pkg/db/v1alpha3"
+	db "github.com/kubeflow/katib/pkg/db/v1alpha3"
+	"github.com/kubeflow/katib/pkg/db/v1alpha3/common"
 	"k8s.io/klog"
 
 	"google.golang.org/grpc"
@@ -20,7 +22,7 @@ const (
 	port = "0.0.0.0:6789"
 )
 
-var dbIf kdb.KatibDBInterface
+var dbIf common.KatibDBInterface
 
 type server struct {
 }
@@ -194,8 +196,11 @@ func (s *server) Check(ctx context.Context, in *health_pb.HealthCheckRequest) (*
 func main() {
 	flag.Parse()
 	var err error
-
-	dbIf, err = kdb.New()
+	dbName := os.Getenv("DB_NAME")
+	if dbName == "" {
+		klog.Fatal("DB_NAME env is not set. Exiting")
+	}
+	dbIf, err = db.NewKatibDBInterface(dbName)
 	if err != nil {
 		klog.Fatalf("Failed to open db connection: %v", err)
 	}
