+++ manifests/v1alpha3/manager/deployment.yaml
@@ -20,7 +20,9 @@ spec:
         image: gcr.io/kubeflow-images-public/katib/v1alpha3/katib-manager
         imagePullPolicy: IfNotPresent
         env:
-          - name: MYSQL_ROOT_PASSWORD
+          - name : DB_NAME
+            value: "mysql"
+          - name: DB_PASSWORD
             valueFrom:
               secretKeyRef:
                 name: katib-db-secrets
