+++ pkg/db/v1alpha3/mysql/init.go
@@ -1,4 +1,4 @@
-package db
+package mysql
 
 import (
 	"fmt"
