+++ pkg/db/v1alpha3/mysql/mysql_test.go
@@ -1,4 +1,4 @@
-package db
+package mysql
 
 import (
 	//	"database/sql"
